Move fetch_, map_, validate_ by endpoint into shared_tasks
