+++ dags/dev_dags/dev_validate_by_mapper_type_to_gdrive.py
@@ -7,10 +7,10 @@ from airflow.models.param import Param
 
 from airflow.providers.google.suite.transfers.local_to_drive import LocalFilesystemToGoogleDriveOperator
 
-from rikolti.dags.utils_by_mapper_type import make_mapper_type_endpoint
-from rikolti.dags.utils_by_mapper_type import fetch_endpoint_task
-from rikolti.dags.utils_by_mapper_type import map_endpoint_task
-from rikolti.dags.utils_by_mapper_type import validate_endpoint_task
+from rikolti.dags.shared_tasks.shared import make_mapper_type_endpoint
+from rikolti.dags.shared_tasks.fetching_tasks import fetch_endpoint_task
+from rikolti.dags.shared_tasks.mapping_tasks import map_endpoint_task
+from rikolti.dags.shared_tasks.mapping_tasks import validate_endpoint_task
 from rikolti.dags.shared_tasks.shared import s3_to_localfilesystem
 
 logger = logging.getLogger("airflow.task")
