+++ dags/shared_tasks/mapping_tasks.py
@@ -1,5 +1,9 @@
 import json
 import os
+import requests
+import sys
+import traceback
+
 from typing import Union, Optional
 
 from airflow.decorators import task, task_group
@@ -8,9 +12,12 @@ from urllib.parse import urlparse
 
 from rikolti.metadata_mapper.lambda_function import map_page
 from rikolti.metadata_mapper.lambda_shepherd import get_mapping_status
+from rikolti.metadata_mapper.map_registry_collections import map_endpoint
+from rikolti.metadata_mapper.map_registry_collections import registry_endpoint
 from rikolti.metadata_mapper.validate_mapping import create_collection_validation_csv
-from rikolti.utils.versions import get_version
 from rikolti.utils.versions import create_mapped_version
+from rikolti.utils.versions import get_mapped_pages
+from rikolti.utils.versions import get_version
 
 @task()
 def create_mapped_version_task(collection, vernacular_page_batches) -> str:
@@ -154,3 +161,116 @@ def mapping_tasks(
     mapped_page_batches = get_mapped_page_filenames_task(mapped_status_batches)
 
     return mapped_page_batches
+
+
+@task()
+def map_endpoint_task(endpoint, fetched_versions, params=None):
+    if not fetched_versions:
+        raise ValueError("No fetched versions provided to map")
+
+    limit = params.get('limit', None) if params else None
+    mapper_job_results = map_endpoint(endpoint, fetched_versions, limit)
+    mapped_versions = {}
+    for mapper_job_result in mapper_job_results:
+        mapped_version = get_version(
+            mapper_job_result['collection_id'],
+            mapper_job_result['mapped_page_paths'][0]
+        )
+        print(
+            "Review mapped data at: https://rikolti-data.s3.us-west-2."
+            f"amazonaws.com/index.html#{mapped_version.rstrip('/')}/data/"
+        )
+        mapped_versions[mapper_job_result['collection_id']] = mapped_version
+
+    return mapped_versions
+
+
+@task()
+def validate_endpoint_task(url, mapped_versions, params=None):
+    if not mapped_versions:
+        raise ValueError("No mapped versions provided to validate")
+
+    limit = params.get('limit', None) if params else None
+
+    response = requests.get(url=url)
+    response.raise_for_status()
+    total = response.json().get('meta', {}).get('total_count', 1)
+    progress = 0
+    if not limit:
+        limit = total
+    limit = int(limit)
+
+    print(f">>> Validating {limit}/{total} collections described at {url}")
+
+    collections = {}
+    data_root = os.environ.get("MAPPED_DATA", "file:///tmp")
+
+    errored_collections = {}
+
+    for collection in registry_endpoint(url):
+        collection_id = collection['collection_id']
+        progress = progress + 1
+        print(f"{collection_id:<6} Validating collection")
+
+        mapped_version = mapped_versions.get(str(collection_id))
+        try:
+            mapped_pages = get_mapped_pages(mapped_version)
+        except (FileNotFoundError, ValueError) as e:
+            print(f"{collection_id:<6}: not mapped yet", file=sys.stderr)
+            errored_collections[collection_id] = e
+            continue
+
+        try:
+            num_rows, version_page = create_collection_validation_csv(
+                collection_id, mapped_pages)
+        except Exception as e:
+            print(f"{collection_id:<6}: {e}", file=sys.stderr)
+            errored_collections[collection_id] = e
+            continue
+
+        collections[collection_id] = {
+            'csv': version_page,
+            'mapped_version': mapped_version,
+            'data_uri': f"{data_root.rstrip('/')}/{version_page}"
+        }
+        print(f"Output {num_rows} rows to {version_page}")
+
+        if limit and progress >= limit:
+            break
+
+    if data_root.startswith('s3'):
+        for collection in collections.values():
+            s3_path = urlparse(collection['data_uri'])
+            bucket = s3_path.netloc
+            print(
+                "Download validation report at: "
+                f"https://{bucket}.s3.amazonaws.com{s3_path.path}"
+            )
+            print(
+                f"Review collection data at: "
+                f"https://{bucket}.s3.us-west-2.amazonaws.com/index.html"
+                f"#{collection['mapped_version'].rstrip('/')}/data/"
+            )
+
+    if len(errored_collections) > 0:
+        print("-" * 60, file=sys.stderr)
+        header = ' Validation Errors '
+        print(f"{header:-^60}", file=sys.stderr)
+        print("-" * 60, file=sys.stderr)
+    for collection_id, e in errored_collections.items():
+        collection_error_header = f" Collection {collection_id}: {e} "
+        print(f"{collection_error_header:>^60}", file=sys.stderr)
+        traceback.print_exception(type(e), e, e.__traceback__, file=sys.stderr)
+        print("<"*60, file=sys.stderr)
+    if len(errored_collections) > 0:
+        print("*" * 60)
+        print(f"{len(errored_collections)} collections encountered validation errors")
+        print(f"please validate manually: {list(errored_collections.keys())}")
+        print("*" * 60)
+
+    if not len(collections):
+        print("-", file=sys.stderr)
+        raise ValueError("No collections successfully validated, exiting.")
+
+    return [collection['csv'] for collection in collections.values()]
+
